Created Filter Sets now selectable for use on HomePage. Default single filters are now also provided.
