+++ client/src/lib/title-mappings.ts
@@ -1,0 +1,14 @@
+/* This is a mapping of each filter's name to the title that should be displayed in the UI */
+export const titleMappings: Record<string, string> = {
+  usSsn: 'US SSN',
+  usDriverLicense: "US Driver's License",
+  usBankNumber: 'US Bank Number',
+  dateTime: 'Date / Time',
+  ipAddress: 'IP Address',
+  creditCard: 'Credit Card',
+  crypto: 'Crypto Wallet',
+  location: 'Location',
+  emailAddress: 'Email Address',
+  phoneNumber: 'Phone Number',
+  person: 'Person',
+};
