+++ client/src/pages/HomePage/RedactOrPrompt.tsx
