+++ client/src/pages/HomePage/WriteBox.tsx
